+++ src/python/magnum/platform/application.h
@@ -88,6 +88,9 @@ template<class T, class Trampoline, class Holder> void application(py::class_<T,
         .def_property_readonly("window_size", &T::windowSize, "Window size")
         .def_property_readonly("framebuffer_size", &T::framebufferSize, "Framebuffer size")
         .def_property_readonly("dpi_scaling", static_cast<Vector2(T::*)() const>(&T::dpiScaling), "DPI scaling")
+        /* Mouse handling */
+        .def_property("cursor", &T::cursor, &T::setCursor, "Cursor type")
+        .def("warp_cursor", &T::warpCursor, "Warp mouse cursor to given coordinates")
         /* Event handlers */
         .def("exit_event", &T::exitEvent, "Exit event")
         .def("viewport_event", &T::viewportEvent, "Viewport event")